+++ fdr-node/src/main/java/com/dwarfeng/fdr/node/configuration/ServiceExceptionMapperConfiguration.java
@@ -25,7 +25,7 @@ public class ServiceExceptionMapperConfiguration {
         destination.put(UnsupportedTriggerTypeException.class, ServiceExceptionCodes.TRIGGER_TYPE_UNSUPPORTED);
         destination.put(TriggerExecutionException.class, ServiceExceptionCodes.TRIGGER_EXECUTION_FAILED);
         destination.put(PointNotExistsException.class, ServiceExceptionCodes.POINT_NOT_EXISTS);
-        destination.put(RecordStoppedException.class, ServiceExceptionCodes.RECORD_HANDLER_STOPPED);
+        destination.put(RecordHandlerStoppedException.class, ServiceExceptionCodes.RECORD_HANDLER_STOPPED);
         destination.put(ConsumeStoppedException.class, ServiceExceptionCodes.CONSUME_HANDLER_STOPPED);
         destination.put(MapperException.class, ServiceExceptionCodes.MAPPER_FAILED);
         destination.put(MapperMakeException.class, ServiceExceptionCodes.MAPPER_MAKE_FAILED);